requirements/test.txt

coverage
django-debug-toolbar
flake8
mock
